backend/api/requirements.txt
fastapi
pyjwt
passlib
bcrypt
sqlalchemy
python-multipart
pymongo[srv]==3.6